requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
Werkzeug==2.2.2
python-docx==0.8.11